requirements.txt
# ══════════════════════════════════════════════════════
# منظومة قوالب WordPress العربية — المتطلبات
# ══════════════════════════════════════════════════════

# ── الأساس ─────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.26.0
aiohttp>=3.9.0

# ── LangGraph + LangChain ──────────────────────────────
langgraph>=0.2.0
langchain>=0.2.0
langchain-anthropic>=0.2.0
langsmith>=0.1.0

# ── نماذج الذكاء الاصطناعي ────────────────────────────
anthropic>=0.35.0
google-generativeai>=0.8.0

# ── قاعدة البيانات ─────────────────────────────────────
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# ── Redis ──────────────────────────────────────────────
redis[hiredis]>=5.0.0
requests>=2.32.0

# ── RAG / Embeddings ───────────────────────────────────
chromadb>=0.5.0
sentence-transformers>=3.0.0
langchain-chroma>=0.1.0
qdrant-client>=1.9.0
openai>=1.0.0

# ── WooCommerce ────────────────────────────────────────
woocommerce>=3.0.0

# ── FastAPI ────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# ── معالجة الصور ──────────────────────────────────────
Pillow>=10.2.0
pillow-avif-plugin>=1.4.0

# ── تحليل النصوص ──────────────────────────────────────
langdetect>=1.0.9
beautifulsoup4>=4.12.0
lxml>=5.1.0

# ── التقارير ───────────────────────────────────────────
jinja2>=3.1.0
markdown>=3.5.0

# ── الجدولة ────────────────────────────────────────────
apscheduler>=3.10.0

# ── المراقبة ───────────────────────────────────────────
structlog>=24.0.0
prometheus-client>=0.19.0

# ── أدوات مساعدة ──────────────────────────────────────
python-magic>=0.4.27
tenacity>=8.2.0      # Retry logic
orjson>=3.9.0        # JSON سريع
